Light sensor Calibrated, Humidity Sensor almost working, enhanced displaying.

+++ Weathermeister/Lighttemphum.h
@@ -47,7 +47,7 @@ extern int8_t dht_gettemperaturehumidity(float *temperature, float *humidity);
 extern int8_t dht_gettemperature(int8_t *temperature);
 extern int8_t dht_gethumidity(int8_t *humidity);
 extern int8_t dht_gettemperaturehumidity(int8_t *temperature, int8_t *humidity);
-extern int Get_Hum();
+extern float Get_Hum();
 extern int Get_Light();
 #endif
